fix: logout 시 header에 session 반영이 안되는 문제
