+++ src/components/header/LogoutButton.tsx
@@ -1,0 +1,32 @@
+'use client';
+
+import Link from 'next/link';
+import type { SessionType } from 'next-auth';
+import { signOut as nextAuthSignOut } from 'next-auth/react';
+import { toast } from '@/components/toast';
+import { getKaKaoLogoutURL } from '@/domain/auth/util';
+
+export default function LogoutForm({
+  sessionType,
+}: {
+  sessionType: SessionType;
+}) {
+  const handleLogout = async () => {
+    await nextAuthSignOut();
+    toast({ message: '로그아웃에 성공했습니다', eventType: 'success' });
+  };
+
+  return sessionType === 'kakao' ? (
+    <KaKaoLogoutLink />
+  ) : (
+    <button onClick={handleLogout}>로그아웃</button>
+  );
+}
+
+function KaKaoLogoutLink() {
+  return (
+    <Link href={getKaKaoLogoutURL()} aria-label='로그아웃'>
+      로그아웃
+    </Link>
+  );
+}
